requirements.txt
streamlit
supabase
langchain
langchain-google-genai
langchain-community
langgraph
python-dotenv
langchain-openai
langchain-groq
fpdf2
python-docx
pdf2docx
pydantic